requirements.txt
asgiref==3.8.1
asyncio==3.4.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
defusedxml==0.8.0rc2
Django==3.2.8
django-admin-cli==0.1.1
django-cors-headers==3.8.0
django-shortuuidfield==0.1.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.2.3
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
oauthlib==3.2.2
Pillow==8.3.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
shortuuid==1.0.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
websockets==12.0
zope.interface==6.4.post2
